analysis_service/requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# LangChain and OpenAI
langchain>=0.1.0
langchain-openai>=0.0.2
langgraph>=0.0.10
openai>=1.0.0
langchain_community>=0.0.10

# Web3 and Ethereum
web3>=6.0.0
hexbytes>=0.3.1

# HTTP and API
requests>=2.31.0
aiohttp>=3.8.0

# Testing
pytest>=7.0.0

# Type checking
typing-extensions>=4.5.0 